Update Redis-template and endDomainEvent

+++ src/main/java/com/hammperpulse/auction/kafka/messaging/dto/AuctionEndedEvent.java
@@ -15,5 +15,5 @@ public class AuctionEndedEvent {
     private LocalDateTime endTime;
     private String result;
     private String winnerName;
-    private BigDecimal winningPrice;
+    private int winningPrice;
 }
